Lint(filename): Tighten up the regex
Follow-up from:
https://github.com/pion/dtls/pull/176#discussion_r366268233

All files must start with a letter. Though this is not mandated by
the Go toolchain, it looks weird having .go files starting with a
number and it's extremely rare in the wild. Additionally the old
regex allowed for filenames to start with an underscore which is
also a bit peculiar.

As long as a file now starts with a letter, it can be followed up by
any combination of letters, numbers and underscores. We could write an
even more pedantic linter, but this is probably good enough.