Add CodeQL based security scanning
This launches a CodeQL scan once a week, on Sunday at 05:23 UTC for the
latest commit on the default branch. The results of the scan will be
available on the repository's Security tab in the Code Scanning alerts.
This is only visible to folks that are org or repo admins (not just
push, admin), or members of a team in the Pion org with the security
manager role.