Set Go version using workflow options
This adds a bit of flexibility to the workflows in that you can now
chose which version(s) of Go you want to run a workflow with.

Since workflow_call inputs can't be an array we introduce the go_prev
input and instead of having supported-go-version-list we now have
previous-go-version. This lets us use both in the matrix config.