Add basic IPv6 support to i386 docker runners
Github runners do not natively support IPv6 [1], which makes it hard to
test IPv6 code, even when external IPv6 connectivity is not needed.

This simple change adds IPv6 support to the docker containers along with
an IPv6 address assigned from a ULA (Unique Local Address) subnet.

There is still no external IPv6 connectivity, but local IPv6 tests can
now be run in the docker containers.

This new IPv6 functionality will be used to re-enable IPv6 tests in the
`mdns` repo, which are currently broken due to lack of IPv6 in docker.

[1] https://github.com/actions/runner-images/issues/668